refactor: Consolidate schemas and remove code duplication
