+++ src/routes/companySchemas.ts
@@ -1,5 +1,1 @@
-import { z } from 'zod'
-
-export const wikidataIdSchema = z.string().regex(/Q\d+/)
-
-export const wikidataIdParamSchema = z.object({ wikidataId: wikidataIdSchema })
+export { wikidataIdSchema, wikidataIdParamSchema } from '../openapi/schemas'
